Trying to get more levels in the glyphs
I was limited to 8 levels before where I'd compute
the levels of each quad of the glyph and match it
to the corresponding quad of levels.

But, it was limited to 8 levels. So I indirected
it through another texture. Now I can have 64 levels.

But, that means I have to check 16 million combinations
of levels to find matching glyphs which takes FOREVER.

I've only done 32 levels so far and that takes like 3-4 mins

And... IT DOESN'T LOOK ANY BETTER!!?!?!

Something must be wrong.... checking in for now